testproject/requirements.txt
Django==3.2.7
graphene-django==2.15.0
pytest-django==4.4.0
django-filter==2.4.0
gunicorn==19.9.0
